+++ examples/resources/nxos_loopback_interface/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_loopback_interface.example "sys/intf/lb-[lo123]"