+++ examples/resources/nxos_loopback_interface/resource.tf
@@ -1,0 +1,5 @@
+resource "nxos_loopback_interface" "example" {
+  interface_id = "lo123"
+  admin_state  = "down"
+  description  = "My Description"
+}