+++ internal/provider/provider.go
@@ -234,6 +234,7 @@ func (p *provider) GetResources(ctx context.Context) (map[string]tfsdk.ResourceT
 		"nxos_feature_pim":            resourceFeaturePIMType{},
 		"nxos_feature_dhcp":           resourceFeatureDHCPType{},
 		"nxos_feature_lldp":           resourceFeatureLLDPType{},
+		"nxos_loopback_interface":     resourceLoopbackInterfaceType{},
 	}, nil
 }
 
@@ -253,6 +254,7 @@ func (p *provider) GetDataSources(ctx context.Context) (map[string]tfsdk.DataSou
 		"nxos_feature_pim":            dataSourceFeaturePIMType{},
 		"nxos_feature_dhcp":           dataSourceFeatureDHCPType{},
 		"nxos_feature_lldp":           dataSourceFeatureLLDPType{},
+		"nxos_loopback_interface":     dataSourceLoopbackInterfaceType{},
 	}, nil
 }