+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add nxos_feature_pim resource and data source
 - Add nxos_feature_dhcp resource and data source
 - Add nxos_feature_lldp resource and data source
+- Add nxos_loopback_interface resource and data source
 
 ## 0.2.3